Initial Implementation of the Public Event system - Implemented public event objective base system. - Implemented public event stats system. - Implemented public event voting system. - Implemented public event entity system (dynamic spawns from script and static spawns from database) - Added basic scripts to start events for Walatiki Temple, The Slaughterdome and Expeditions.
In addition, a few changes to matches were made
- Updated matching queues to allow for cross faction groups (except for Adventures and Open Arena)
- Implemented match leave
- Implemented teleport to match
- Implemented temporary factions, this is used when joining cross faction matches
- Initial PvP match implementation

+++ Source/NexusForever.Game/Entity/VehicleEntity.cs
@@ -189,7 +189,7 @@ namespace NexusForever.Game.Entity
                 EnqueueToVisible(new ServerEntityFaction
                 {
                     UnitId  = Guid,
-                    Faction = (ushort)player.Faction1
+                    Faction = player.Faction1
                 }, true);
 
                 Faction1 = player.Faction1;
